# apps/agents-server/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agents Server API",
        "version": "1.0.0",
        "description": "API documentation for the Agents Server"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current Server"
        }
    ],
    "paths": {
...
        "/api/agents/{agentName}": {
            "get": {
                "summary": "Get agent details",
                "tags": ["Agents"],
                "parameters": [
                    {
                        "name": "agentName",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent details"
                    }
                }
            }
        },
...
    },
...
}